:hammer: Refactoring - more components
